% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NgramTools.R
\name{ngram_count}
\alias{ngram_count}
\title{Count word frequency from Google Ngram data.}
\usage{
ngram_count(data, years, words = NULL, pattern = NULL)
}
\arguments{
\item{data}{A \code{data.table} like the demodata (see \link{GoogleNgramData}).}

\item{years}{A numeric vector of years to be included.}

\item{words}{[Option 1] Word strings (\code{NULL}; a single word; a vector of words).}

\item{pattern}{[Option 2] Pattern of regular expression (see \code{\link[stringr:str_subset]{str_subset}}).}
}
\value{
A numeric vector of word counts for all specified years.
}
\description{
Count word frequency from Google Ngram data.
}
\examples{
ngram_count(ngram.v3.chi_TAG,
            years=1949:2019,
            pattern="^独特_|^独一无二_")

years = 1949:2019
pattern = "^独特_|^独一无二_"
data = data.table(
  year = years,
  total.words = total.v3[year \%in\% years]$words.chi,
  freq = ngram_count(ngram.v3.chi_TAG, years, pattern=pattern)
)
data[, prop := freq / total.words]
data

}
